requirements.txt
# --- Base ---
flask
flask-cors
python-dotenv
gunicorn

# --- Langchain ---
langchain
langchain-core
langchain-community
langchain-groq

# --- Google Drive/Cloud ---
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth

# --- Embeddings/Vector Stores ---
faiss-cpu

# --- Data Models ---
pydantic

# --- Utilitaires ---
requests

# --- PostgreSQL ---
supabase==1.0.3

pdfplumber 
python-docx




